# src/backend/services/basket.api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a rest api for basket which saves items to redis server",
        "title": "Basket API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/items": {
            "post": {
                "security": [
                    {
                        "OAuth": []
                    }
                ],
                "description": "add by json new CustomerBasket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomerBasket"
                ],
                "summary": "Add a CustomerBasket",
                "parameters": [
                    {
                        "description": "Add CustomerBasket",
                        "name": "CustomerBasket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.CustomerBasket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.CustomerBasket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth": []
                    }
                ],
                "description": "Get CustomerBasket by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomerBasket"
                ],
                "summary": "Gets a CustomerBasket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CustomerBasket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.CustomerBasket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth": []
                    }
                ],
                "description": "Deletes CustomerBasket by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomerBasket"
                ],
                "summary": "Deletes a CustomerBasket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CustomerBasket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BasketItem": {
            "type": "object",
            "properties": {
                "foodId": {
                    "type": "string"
                },
                "foodName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "oldUnitPrice": {
                    "type": "number"
                },
                "picture": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "unitPrice": {
                    "type": "number"
                }
            }
        },
        "models.CustomerBasket": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BasketItem"
                    }
                }
            }
        },
        "models.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "{{.AuthUrl}}",
            "scopes": {
                "basket-api": " Access to basket-api"
            }
        }
    }
}